Split worker threads into Blueprints v1- and v2- specific
